+++ player/player.py
@@ -1,7 +1,7 @@
 import sys
 import os
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(__file__))))
-
+import random
 from datetime import date
 import json
 import time
@@ -10,9 +10,11 @@ class Player:
     def __init__(self):
         self.date = date.today()
         self.score = 0
+        self.lost_body_parts = []
         try:
             with open("player/bodyParts.json", 'r') as content:
                 self.bodyParts = json.load(content)["bodyParts"]
+                self.probability = json.load(content)["probability"]
         except FileNotFoundError:
             print("We didn't find any of your body parts, you died...")
             # file not found
@@ -52,9 +54,32 @@ class Player:
         record["Total Player"] += 1
         record[str(total_num+1)] = {"Name": self.name,
                                     "Date": str(self.date),
-                                    "Score": self.score}
+                                    "Score": self.score,
+                                   "Game record": {"Game 1": None,# win/lose: 1/0
+                                                    "Game 2": None,
+                                                    "Game 3": None,
+                                                    "Game 4": None},
+                                    "Lost body parts": []}
         with open("player/playingRecord.json", 'w') as file:
-            json.dump(record, file, indent = 4)
+            json.dump(self.record, file, indent = 4)
+    
+    def choose_body_part(self):
+        r = random.randint(1, 1000)
+        for i in range(len(self.probability)-1,0):
+            if r <= self.probability[i]:
+                if self.probability[i] in self.lost_body_parts:
+                    self.choose_body_part()
+                    break
+                self.lost_body_parts.append(self.bodyParts[i])
+                return self.bodyParts[i]
+    
+    def lose(self, body_part):
+        self.lost_body_parts.append(body_part)
 
 
+def clear_all_playing_records():
+    with open("player/playingRecord.json", 'w') as file:
+        json.dump({"Total Player": 0}, file, indent = 4)
+
+#clear_all_playing_records()