+++ changelog.md
@@ -2,3 +2,18 @@
 <br><br>
 
 ## Change Log
+
+1.1.# 
+- finish *player* folder
+    + bodyParts.json
+    + player.py
+    + playingRecord.json
+- finish `main.py` in main folder
+    + can now import all other four mini-games
+    + creat four objects inside `Game` class
+    + can run the four games inside folder
+    + import `player.py` from *player* folder
+    + creat a player object to read player information, store player information and store playing record
+
+1.2.#
+- finish game 2